Adapt unit tests
