+++ src/actions/forging.js
@@ -1,26 +1,33 @@
 import actionTypes from '../constants/actions';
 import { getForgedBlocks, getForgedStats } from '../utils/api/forging';
 
-export const updateForgedBlocks = data => ({
+export const forgedBlocksUpdated = data => ({
   data,
   type: actionTypes.forgedBlocksUpdated,
 });
 
-export const fetchAndUpdateForgedBlocks = (activePeer, limit, offset, generatorPublicKey) =>
+export const fetchAndUpdateForgedBlocks = ({ activePeer, limit, offset, generatorPublicKey }) =>
   (dispatch) => {
-    getForgedBlocks(activePeer, limit, offset, generatorPublicKey).then((response) => {
-      dispatch(updateForgedBlocks(response.blocks));
-    });
+    getForgedBlocks(activePeer, limit, offset, generatorPublicKey)
+    .then(response =>
+      dispatch(forgedBlocksUpdated(response.blocks)),
+    ).catch(() =>
+      dispatch(forgedBlocksUpdated()),
+    );
   };
 
-export const updateForgingStats = data => ({
+export const forgingStatsUpdated = data => ({
   data,
   type: actionTypes.forgingStatsUpdated,
 });
 
-export const fetchAndUpdateForgedStats = (activePeer, key, startMoment, generatorPublicKey) =>
+export const fetchAndUpdateForgedStats = ({ activePeer, key, startMoment, generatorPublicKey }) =>
   (dispatch) => {
-    getForgedStats(activePeer, startMoment, generatorPublicKey).then((response) => {
-      dispatch(updateForgingStats({ [key]: response.forged }));
-    });
+    getForgedStats(activePeer, startMoment, generatorPublicKey)
+    .then(response =>
+      dispatch(forgingStatsUpdated({ [key]: response.forged })),
+    )
+    .catch(() =>
+      dispatch(forgingStatsUpdated()),
+    );
   };
